+++ src/aui-progressbar/HISTORY.md
@@ -1,2 +1,4 @@
 aui-progressbar
 ========
+
+	* #AUI-2058 AUI Progress Bar should use percentage instead of px when horizontal to make it responsive